refactor(compartment-mapper)!: Cleanly separate StaticModuleRecord dependency
