+++ packages/compartment-mapper/src/import-archive.js
@@ -3,13 +3,22 @@
 
 import { readZip } from '@endo/zip';
 import { assemble } from './assemble.js';
-import { parserForLanguage } from './parse.js';
+import { parsePreCjs } from './parse-pre-cjs.js';
+import { parseJson } from './parse-json.js';
+import { parsePreMjs } from './parse-pre-mjs.js';
 import * as json from './json.js';
 
 // q as in quote for strings in error messages.
 const q = JSON.stringify;
 
 const textDecoder = new TextDecoder();
+
+/** @type {Record<string, ParseFn>} */
+export const parserForLanguage = {
+  precjs: parsePreCjs,
+  premjs: parsePreMjs,
+  json: parseJson,
+};
 
 /**
  * @callback ArchiveImportHookMaker
@@ -103,6 +112,7 @@ export const parseArchive = async (archiveBytes, archiveLocation) => {
     );
     const compartment = assemble(compartmentMap, {
       makeImportHook,
+      parserForLanguage,
       globals,
       globalLexicals,
       modules,
