+++ packages/compartment-mapper/src/import.js
@@ -6,7 +6,17 @@ import { compartmentMapForNodeModules } from './node-modules.js';
 import { search } from './search.js';
 import { assemble } from './assemble.js';
 import { makeImportHookMaker } from './import-hook.js';
+import { parseJson } from './parse-json.js';
+import { parseCjs } from './parse-cjs.js';
+import { parseMjs } from './parse-mjs.js';
 import * as json from './json.js';
+
+/** @type {Record<string, ParseFn>} */
+export const parserForLanguage = {
+  mjs: parseMjs,
+  cjs: parseCjs,
+  json: parseJson,
+};
 
 /**
  * @param {ReadFn} read
@@ -51,6 +61,7 @@ export const loadLocation = async (read, moduleLocation, options) => {
     const makeImportHook = makeImportHookMaker(read, packageLocation);
     const compartment = assemble(compartmentMap, {
       makeImportHook,
+      parserForLanguage,
       globals,
       globalLexicals,
       modules,
